fix: handles already inlined content
Fixes #26

Inliner was trying to resolve and request `data:` urls. Now it checks to see if it's already inlined.

I also updated the whole `.get` method to make smarter use of the caching. Originally the cache is only populated when the request has completed, but if the same image is on the page 20 times, the request would be made 20 times. Now, we cache the promise, so the request is only placed once, and returns a re-usable promise if it's already in progress or resolved.